+++ crates/bevy_core/src/lib.rs
@@ -6,7 +6,7 @@ mod name;
 mod serde;
 mod task_pool_options;
 
-use bevy_ecs::system::Resource;
+use bevy_ecs::system::{ResMut, Resource};
 pub use bytemuck::{bytes_of, cast_slice, Pod, Zeroable};
 pub use name::*;
 pub use task_pool_options::*;
@@ -53,6 +53,7 @@ impl Plugin for CorePlugin {
         register_math_types(app);
 
         app.init_resource::<FrameCount>();
+        app.add_system(update_frame_count);
     }
 }
 
@@ -108,6 +109,10 @@ fn register_math_types(app: &mut App) {
 #[derive(Default, Resource, Clone, Copy)]
 pub struct FrameCount(pub u32);
 
+fn update_frame_count(mut frame_count: ResMut<FrameCount>) {
+    frame_count.0 = frame_count.0.wrapping_add(1);
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -144,5 +149,15 @@ mod tests {
         async_rx.try_recv().unwrap();
         compute_rx.try_recv().unwrap();
         io_rx.try_recv().unwrap();
+    }
+
+    #[test]
+    fn frame_counter_update() {
+        let mut app = App::new();
+        app.add_plugin(CorePlugin::default());
+        app.update();
+
+        let frame_count = app.world.resource::<FrameCount>();
+        assert_eq!(1, frame_count.0);
     }
 }
