+++ crates/bevy_render/src/globals.rs
@@ -18,10 +18,15 @@ impl Plugin for GlobalsPlugin {
             render_app
                 .init_resource::<GlobalsBuffer>()
                 .init_resource::<Time>()
+                .add_system_to_stage(RenderStage::Extract, extract_frame_count)
                 .add_system_to_stage(RenderStage::Extract, extract_time)
                 .add_system_to_stage(RenderStage::Prepare, prepare_globals_buffer);
         }
     }
+}
+
+fn extract_frame_count(mut commands: Commands, frame_count: Extract<Res<FrameCount>>) {
+    commands.insert_resource(**frame_count);
 }
 
 fn extract_time(mut commands: Commands, time: Extract<Res<Time>>) {
