+++ crates/bevy_render/src/lib.rs
@@ -19,7 +19,6 @@ mod spatial_bundle;
 pub mod texture;
 pub mod view;
 
-use bevy_core::FrameCount;
 use bevy_hierarchy::ValidParentCheckPlugin;
 pub use extract_param::Extract;
 
@@ -333,8 +332,7 @@ impl Plugin for RenderPlugin {
             .add_plugin(CameraPlugin)
             .add_plugin(ViewPlugin)
             .add_plugin(MeshPlugin)
-            .add_plugin(GlobalsPlugin)
-            .add_plugin(FrameCountPlugin);
+            .add_plugin(GlobalsPlugin);
     }
 }
 
@@ -367,23 +365,4 @@ fn extract(app_world: &mut World, render_app: &mut App) {
     // so that in future, pipelining will be able to do this too without any code relying on it.
     // see <https://github.com/bevyengine/bevy/issues/5082>
     extract.apply_buffers(running_world);
-}
-
-pub struct FrameCountPlugin;
-impl Plugin for FrameCountPlugin {
-    fn build(&self, app: &mut bevy_app::App) {
-        app.add_system(update_frame_count);
-
-        if let Ok(render_app) = app.get_sub_app_mut(RenderApp) {
-            render_app.add_system_to_stage(RenderStage::Extract, extract_frame_count);
-        }
-    }
-}
-
-fn update_frame_count(mut frame_count: ResMut<FrameCount>) {
-    frame_count.0 = frame_count.0.wrapping_add(1);
-}
-
-fn extract_frame_count(mut commands: Commands, frame_count: Extract<Res<FrameCount>>) {
-    commands.insert_resource(**frame_count);
 }
